refactor: select menu handler using Collection()
Closes #333
