+++ main.js
@@ -83,10 +83,11 @@ async function handleDatabaseError(err) {
 
 data.guild.instance.on('error', handleDatabaseError);
 
-/** @type {Client & {commands: Collection, buttons: Collection, music: Node}} */
+/** @type {Client & {commands: Collection, buttons: Collection, selects: Collection, music: Node}} */
 const bot = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_PRESENCES, Intents.FLAGS.GUILD_MEMBERS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILD_VOICE_STATES] });
 bot.commands = new Collection();
 bot.buttons = new Collection();
+bot.selects = new Collection();
 bot.music = new Node({
 	connection: {
 		host: lavalink.host,
